refactor: replace `Predicate` w/ `&[Expr]` in querier internals (#8465)
* refactor: replace `Predicate` w/ `&[Expr]` in querier internals

First step towards #8097. This replaces most internal usages of
`Predicate` with the more appropriate `&[Expr]` within the querier code.

This is also triggered by #8443 because the new ingester protocol shall
not use `Predicate` anymore.

Note that the querier still uses `Predicate` for a few interfaces. These
will be fixed later:

- the current ingester RPC version
- chunk pruning
- `QuerierNamespace::chunks`

* fix: docs