feat(gossip): topics & topic sets
Adds an (unused) Topic and TopicSet type to the gossip crate.

A Topic is a number from 0 to 63 inclusive that uniquely identifies a
"type" of application payload. The TopicSet efficiently encodes the set
of topics a node is interested in as a 64-bit wide bitmap, with each
Topic encoded as a single bit, shifted right N times for the topic
value.

This allows for very cheap "is this node interested" set evaluation and
space-efficient propagation of interest sets during PEX.