refactor(gossip): rename GossipMessage
Now there's a Topic, there's no need for a giant "all message types"
enum.

As part of this shift, the gossip_message::GossipMessage used for schema
gossiping is sounding overly generic. This commit changes the name to
schema_message::SchemaMessage and updates the code.

This is a backwards-compatible change (and if anything goes wrong, the
"old" routers simply log a warning if a message is unreadable).