fix: do not panic in Influx selectors (#8513)
Selectors that are value-based (like "min" and "max") and that also
track "other" fields (so things apart from the timestamp and the value),
panicked when the saw the same min/max value in different batches but
the timestamp did was NOT smaller than the previously selected value. An
example would be:

Batch | time | value | other
----- | ---- | ----- | ------
1     | 0    | 0     | 0
2     | 0    | 0     | 1

This adds the fix and a test.

Closes https://github.com/influxdata/idpe/issues/17968 .

Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>