perf(anti_entropy): add AntiEntropyActor for MST
Separate the management of the Merkle Search Tree state into an actor
to manage concurrent access.

This moves hashing and tree updates off of the hot request path, and
into an asynchronous background process, practically eliminating
overhead for maintaining the MST structure.

This decoupling will allow convergence runs between peers to proceed
without causing contention on the lock in the request hot path.