perf: gossip new parquet file persistence events
Configures the ingesters to gossip parquet file persistence
notifications to listening peers.

Because there's no current listeners, this will be a no-op - the gossip
subsystem will not send the messages as no peer will be registered as
interested in them, so the frame rate metrics, etc, will not change
until something registers as interested in the new topic.