feat(gossip): peer socket address lookup
This allows a caller to obtain the socket address of a peer identity,
and therefore the address from which a message came from.

I think the use cases of this are pretty slim (which is why I purposely
kept the concept of peers out of the handle) but it definitely is useful
as a primitive service discovery directory.