fix(influxql): FILL(linear) for selectors (#8396)
* fix(influxql): FILL(linear) for selectors

Ensure that selector functions such as FIRST, LAST, MIN and MAX can
use LINEAR filling in the same way as influxdb 1.8.

* chore: review suggestions

Apply suggestions from the review. This adds more tests and support
for interpolation in SQL.

* fix: lint

* fix: lint

* chore: buffered input for struct arrays

Ensure that for linear interpolation the buffered input of a struct
field ensures that buffering only stops when there is a non-null
struct containing a non-null value.

* fix: integration test

* fix(iox_query): make clippy happy

---------

Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>