fix(router): health probe livelock / missing probe
When an upstream ingester goes offline, the "circuit breaker" detects it
as unhealthy, and prevents further requests being sent to it.
Periodically a small number of requests are allowed ("probe requests")
to check for recovery.

If a write request is selected as a "probe request", it SHOULD be sent -
a limited number writes are selected as probes, and enough have to be
successful to drive recovery. If no probes are ever sent/successful, the
upstream will never be marked as healthy.

Additionally the RPC handler applies an optimisation: if the number of
ingesters selected to service a write is less than the number needed to
successfully reach the desired replication factor, no requests are sent
and an error is returned immediately, preventing unnecessary system load
for writes that would never succeed.

This optimisation conflicts with the probe request requirement when a
replication factor of >= 2 is specified:

   * All ingesters are offline
   * Write comes in
   * UpstreamSnapshot is populated with a probe request for 1 ingester
     only - no other healthy candidate ingesters exist.
   * Optimisation applied: 1 probe candidate < 2 needed for replication

This results in a probe request never being sent, and in turn, never
allowing further requests to the recovered upstream.

This fix changes the optimisation, applying it only when there are no
probes in the candidate ingester list - the write will always fail, but
it will drive detection of recovered ingesters and maintain liveness of
the system.