feat(querier): convert timezone sent from ingester (#8769)
* feat(querier): convert timezone sent from ingester

In order to facilitate the change of default timezone from None to
UTC make the querier able to convert the timezone sent from the
ingester into its preferred type. This can convert from None to UTC
or UTC to None and should allow the interaction between ingesters
and queriers with differing settings for the default timezone.

To allow testing of both conversions, the type checking has been
made more liberal when converting an arrow schema to an IOx one.

* fix: fmt

* fix: lint

---------

Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>