feat: Add authenticating ID and user ID to request logging (#24474)
To allow rudimentary security auditing of logs,
add the authenticating ID and the user ID when
possible to the request logs. When a request is
authorized for V1 or V2 API, store the authorizer
object to be used by the logger up the call stack.

closes https://github.com/influxdata/influxdb/issues/24473

(cherry picked from commit 13cf6be49ec6c23ac99cd26cc745814712985e0d)

closes https://github.com/influxdata/influxdb/issues/24475