feat: Add paths module to influxdb3_write
This commit introduces 3 new types in the paths module for the
influxdb3_write crate. They are:

- ParquetFilePath
- CatalogFilePath
- SegmentFilePath

Each of these corresponds to an object store path and also on disk path
that we can use to address the needed files in a consistent way and not
need to have path construction be duplicated to address these files.

These types also Deref/AsRef to the Path type so that they can be used
in places that expect the type such as various std::fs methods and so
that we can use methods like `exist()` without needing to implement them
for each type as they are thin wrappers around PathBuf.

This commit adds some tests to make sure that the path construction
works as intended and also updates the `wal.rs` file to use the new
`SegmentFilePath` instead of just a `PathBuf`. Currently it assumes that
the prefix is just the dir handed to it.

Closes: #24578