feat: add the api_v3_query_sql method to influxdb3 Client
The api_v3_query_sql method was added to the influxdb3 Client, which
allows caller to compose a request to the respective API.

Similar to the write_lp API, this uses a builder approach to composing
the request. The handling of the response was kept naive and just gives
back the bytes returned. We may improve this in future once the format
handling in the API is nailed down and we have JSON lines supported.